Restore ActiveSupport::Duration's constructor
Some users are using ActiveSupport::Duration directly (instead of
through Numeric core extensions), so it's rude to destroy the backwards
compatibility in the constructor, when it isn't necessary.

We're removing the current constructor, because it would be a pain to
maintain both the options hash, and ActiveSupport's constructor. It
technically is a backwards incompatibility, but the gem is only 10 days
old, so it's very unlikely that someone was already using as-duration in
a way that they're instantiating AS::Duration directly. It's also
acceptable because the version is still 0.x, so semantic versioning is
not strictly required, and bumping the minor version should be good
enough.

Fixes #3.